move master-client to master folder
